requirements.txt
Flask==3.1.0
logging==0.4.9.6
argparse==1.4.0
gitpython==3.1.41
